+++ samples/material_sandbox.cpp
@@ -885,15 +885,12 @@ static void gui(filament::Engine* engine, filament::View*) {
             ImGui::SliderFloat("Polygon Offset Scale", &params.polygonOffsetSlope, 0.0f, 10.0f);
             ImGui::SliderFloat("Polygon Offset Constant", &params.polygonOffsetConstant, 0.0f, 10.0f);
 
-            bool* lispsm;
-            if (debug.getPropertyAddress<bool>("d.shadowmap.lispsm", &lispsm)) {
-                ImGui::Checkbox("Enable LiSPSM", lispsm);
-                if (*lispsm) {
-                    ImGui::SliderFloat("dzn",
-                            debug.getPropertyAddress<float>("d.shadowmap.dzn"), 0.0f, 1.0f);
-                    ImGui::SliderFloat("dzf",
-                            debug.getPropertyAddress<float>("d.shadowmap.dzf"),-1.0f, 0.0f);
-                }
+            ImGui::Checkbox("Enable LiSPSM", &params.lispsm);
+            if (params.lispsm) {
+                ImGui::SliderFloat("dzn",
+                        debug.getPropertyAddress<float>("d.shadowmap.dzn"), 0.0f, 1.0f);
+                ImGui::SliderFloat("dzf",
+                        debug.getPropertyAddress<float>("d.shadowmap.dzf"),-1.0f, 0.0f);
             }
             ImGui::Unindent();
         }
@@ -942,6 +939,7 @@ static void gui(filament::Engine* engine, filament::View*) {
 
     LightManager::ShadowOptions options = lcm.getShadowOptions(lightInstance);
     options.stable = params.stableShadowMap;
+    options.lispsm = params.lispsm;
     options.normalBias = params.normalBias;
     options.constantBias = params.constantBias;
     options.polygonOffsetConstant = params.polygonOffsetConstant;