+++ RELEASE_NOTES.md
@@ -6,6 +6,7 @@ A new header is inserted each time a *tag* is created.
 ## main branch
 
 - engine: LiSPSM is now a user settable option
+- engine: get the morph target buffer to the given primitive
 - Java: Fix TransformManager.getChildren()
 - Metal: newer devices are no longer limited to 16 samplers per Material.
 - gltfio: fix interpretation of occlusion strength